Refactor assertions
